refactor: (피드백반영) 번호 생성로직 Number VO 할당
